banque {lycee}	R Documentation

banque /données banquaires

Description

Une agence de banque dispose d'un fichier de données sur 810 clients (ces données ont été récupérées dans le package ade4). Les variables sont qualitatives. Les banques collectent de nombreuses données sur leurs clients, pour mieux les connaître avant de leurs proposer les services les mieux adaptés (crédit confiance, épargne logement, crédit à la consommation etc...).

Quelle est la proportion d'agriculteurs, de retraités, d'étudiants... parmi leurs clients ? Ce client pourra-t-il rembourser ce prêt ? Qui est le plus susceptible de prendre une carte bleue ? Qui risque d'être interdit banquaire ?...

Usage

data(banque)
lire(banque)

Format

Un tableau de 810 lignes et 21 colonnes.

csp

catégorie socio-professionnelle du client agric (agriculteur exploitant), artis (artisant, commercant, chef), cadsup (cadre supérieur), inter (professions intermédiaire), emplo (employés), ouvri (ouvrier), retra (retraité), inact (inactif) etudi (étudiant) .

duree

durée des relations avec le clients dm2 (<2ans),d24 (entre 2 et 4 ans), d48 (entre 4 et 8 ans), d812 (entre 8 et 12 ans), dp12 plus de 12 ans

oppo

non (n'a pas fait opposition à un chèque), oui (a fait opposition).

age

age du clientai25 (entre 18 et 25 ans), ai35 (entre 25 et 35 ans), ai45 (entre 45 et 55 ans), ai55 (entre 45 et 55 ans) ai75 (entre 55 et 75 ans).

sexe

sexe du client.

interdit

oui (est interdit banquaire) non (n'est pas .)

cableue

oui (possède une carte bleue) non (ne possède pas. )

assurvi

oui(a une assurance vie),non (n'a pas .)

soldevu

solde des comptes à vue n1 (débiteur <4000), n2 (débiteur>4000), p1 (créditeur <4000), p2 (créditeur entre 4000 et 12000), p3 (créditeur entre 12000 et 20000), p4 (créditeur >20000)

eparlog

montant des épargnes logement fai (<20000), for (>20000), nul (nulle.)

eparliv

montant des épargnes sur livret fai (<20000), for (>20000), nul (nulle)

credhab

titulaire de crédit habitat non, oui

credcon

montant des crédits à la consommation fai(<20000) for(>20000), nul(aucun.)

versesp

versement en espèce non oui

retresp

retrait en espèce fai(<2000), for(>5000) moy(entre 2000 et 5000.)

remiche

montant des remises de chèques fai(<5000), for(>10000), moy(de 5000 à 10000), nul

preltre

montant des prélèvements du trésor fai(<1000), moy(>1000), nul.

prelfin

montant des prélèvements des organismes fai (<1000), moy(>1000), nul.

viredeb

montants des virements au débit fai(<2500), for(>5000), moy (entre 2500 et 5000), nul.

virecre

montant des virments au crédit fai(<5000), for(>10000), moy(entre 5000 et 10000), nul.

porttit

estimation des portefeuilles titres fai(<20000), for(>100000), moy(entre 20000 et 100000), nul.

Examples


data(banque)
lire(banque)
camembert(credcon)
diagrammeenbarre(age)


[Package lycee version 1.0 Index]