bebe {lycee}	R Documentation

bebe/données médicales

Description

Une chercheuse a mené une étude sur les naissances en milieu hospitalier au Quebec en 2000. elle a choisi arbitrairement un hopital général d'un centre urbain et a consulté les dossiers des enfants nés au cours du mois de mai. Elle y a trouvé les informations suivantes sur le sexe, la masse et la longueur des nouveau-nés à la naissance. Elle se demande par exemple si les distributions des longueurs et poids des nouveau-nés dépend du sexe. Elle veut savoir à partir de quel poids on peut dire qu'un bébé est anormalement gros ou maigre etc...

Usage

data(bebe)
lire(bebe)

Format

Un tableau contenant 1000 observations et 4 variables.

ind

numero du bébé

sexe

sexe du bébé

long

longueur du bébé

poids

poids du bébé

Examples

data(bebe) 
lire(bebe)
hist(long)
moyenne(long)
mediane(long)
ecarttype(long)
hist(poids)
boitesparalleles(poids,sexe)

[Package lycee version 1.0 Index]