cholesterol {lycee}	R Documentation

cholesterol/données de type médical.

Description

On a relevé le taux de cholestérol de 2000 patients, fumeurs ou non fumeurs. On s'attendait à une distribution normale. On explique la non normalité des données par la présence de fumeurs parmi les patients. Un étude plus poussée montrera que les fumeurs ont un taux de cholestérol plus élevé que les non-fumeurs. Pour étudier le taux de cholestérol, on devra séparer les fumeurs et les non-fumeurs.

Usage

data(cholesterol)
lire(cholesterol)

Format

Un tableau contenant 2000 observations sur 2 variables.

taux

une variable numérique

fumeur

une variable qualitative prenant les valeurs oui ou non

Examples


data(cholesterol)
lire(cholesterol)
hist(taux)
mean(taux)
mediane(taux)
boitesparalleles(taux,fumeur)
histogramme(selection(taux,fumeur,"oui"))

[Package lycee version 1.0 Index]