liretableautxt {lycee}	R Documentation

Fonction qui lit un tableau de données format txt.

Description

Cette fonction permet de lire un tableau de données au format txt . Les observations sont séparées par des espaces et chaque colonne a un titre. Pour lire un tableau excell, on doit d'abord en faire un fichier txt. Le tableau doit-être dans le répertoire courant. Les nombres décimaux sont écris avec des virgules.

Usage


nomdufichier=liretableautxt('nomdufichier.txt')

Arguments

d 	 chemin absolu suivi du nom du tableau.

[Package lycee version 1.0 Index]