loinormale {lycee}	R Documentation

Function qui calcule, pour une variable qui suit une loi normale,la probabilité d'être inférieur à ou supérieur à un nombre.

Description

La loi normale est parfaitement connue par les statisticiens. Si on connait sa moyenne et son écart-type, la distribution de la variable est parfaitement connue.

Usage


normale(x,m,s)

Arguments

x 	 la borne de l'intervalle

m 	 la moyenne de la variable étudiée

s 	 l'écart-type de la variable étudiée

Examples


loinormale(4,3,1)


[Package lycee version 1.0 Index]