lycee-package {lycee}	R Documentation

Utilisation du logiciel R au collège et au lycée

Description

Le premier objectif de ce package est de permettre à des collégiens ou lycéeens francophones de découvrir les statistiques et les probabilités. En utilisant le logiciel R, ils peuvent simuler puis décrire des jeux de données de grandes tailles. Ils découvrent ainsi l'approche fréquentielle des probabilités.

Le deuxième objectif est de présenter aux élèves le métier de statisticien et leur montrer l'utilité des mathématiques dans des cadres trés concrets . En effet, les données proposées dans le package sont "inspirées" de données réelles. Ils peuvent ainsi s'initier à un logiciel professionnel utilisé dans les centres de recherche et les entreprises et travailler sur de "vraies" études.

On peut également travailler sur des tableaux de données (excell par exemple) préalablement convertis en format txt.

Author(s)

Irem d'Orléans
jean-pierre.lamarche@univ-orleans.fr
sophie.jacquot@univ-orleans.fr

Examples

 d<-pileface(100)
 frequence(d)
diagrammeenbarre(d)
data(entreprise)
lire(entreprise)
histogramme(ratio)
boitesparalleles(ratio,etat)


[Package lycee version 1.0 Index]