m {lycee}	R Documentation

message codé/texte

Description

Message codé par des pirates dans une nouvelle d'Edgar Poe, décrivant l'emplacement d'un trésor.

Usage

data(m)

Format

A data frame with 229 observations on the following variable.

texte

a factor with levels a c d e f g i j k l m n o p q r s u y z

Examples


data(m)
lire(m)

[Package lycee version 1.0 Index]