nuagedepoints {lycee}	R Documentation

Fonction qui trace le nuage de points représentant deux variables quantitatives.

Description

On dispose de deux variables quantitatives x et y, et on représente les points (x,y) dans le plan.

Usage


nuagedepoints(x,y,'black')

Arguments

x 	 une variable quantitative.

y 	 une variable quantitative.

couleur 	 une variable donnant la couleur de chaque point, ou bien une couleur 'black','red','green'.... qui sera celle de tous les points

Examples


data(urbanisme)
lire(urbanisme)
nuagedepoints(tauxnatalite,tauxurbanisation,'red')
regression(tauxnatalite,tauxurbanisation)
data(kystediscrimin1)
lire(kystediscrimin1)
nuagedepoints(disc1,disc2,coul)


[Package lycee version 1.0 Index]