random {lycee}	R Documentation

Fonction qui donne un nombre au hasard dans un intervalle [a,b].

Description

Cette fonction qui est en général présente dans les calculettes, peut permettre de simuler des phénomènes aléatoires comme les jeux de pile ou face ou de dés.

Usage


random(a, b)

Arguments

a 	 un nombre

b 	 un nombre plus grand que a

Examples


r<-random(0,1)
if (r<0.5){piece="pile"} else {piece="face"}
piece

hist(r)
mediane(r)
moyenne(r)
variance(r)
ecarttype(r)


[Package lycee version 1.0 Index]