simuleexponentielle {lycee}	R Documentation

Function qui simule une distribution exponentielle et trace son histogramme.

Description

La distribution exponentielle est un peu moins fréquente dans la nature que la distribution normale, mais elle est également très présente et parfaitement maitrisée par les statisticiens. On la trouve notamment en fiabilité, où l'on modélise les durée de vie des appareils pour prévoir les risques de panne. On l'utilise aussi pour modéliser les temps d'attente (embouteillages, queues au guichet, réseaux téléphoniques…) afin d'améliorer les services. La loi exponentielle est aussi trés présente en météo (niveaux d’eau d’un fleuve et risques de crues, pluviométrie...) et dans bien d'autres phénomènes.

Usage


d<-simuleexponentielle(n,l)

Arguments

n 	 le nombre de simulations

l 	 inverse de la moyenne

Examples


simuleexponentielle(1000,1/4)


[Package lycee version 1.0 Index]