simulenormale {lycee}	R Documentation

Function qui simule une distribution normale et trace son histogramme.

Description

La loi normale est parfaitement connue par les statisticiens c’est le modèle de loi statistique le plus courant. Dans des domaines infiniment variés, la forme des histogrammes est en "cloche" et la boite à moustache est symétrique. La capacité pulmonaire chez l’adulte suit une loi normale de moyenne 5l et d’écart-type 0,5l. Le test de QI est censé mesurer une loi gaussienne de moyenne 100 et d’écart-type 15, le taux de cholestérol, la pression artérielle, ... les variations du prix des actions, le cout des réparations d'une voiture ayant subi un choc frontal à 50 km/h…

Usage


simulenormale(n,m,s)

Arguments

n 	 le nombre de simulations

m 	 la moyenne des simulations

s 	 l'écart-type des simulations

Examples


simulenormale(1000,4,1)


[Package lycee version 1.0 Index]