simulepoisson {lycee}	R Documentation

Function qui simule une distribution de poisson et trace son histogramme.

Description

Le domaine d'application de la loi de Poisson a été longtemps limité à celui des événements rares comme les suicides d'enfants, les arrivées de bateaux dans un port ou les accidents dûs aux coups de pied de cheval dans les armées... Mais depuis quelques décennies son champ d'application s'est considérablement élargi. On la retrouve dans d'innombrables processus de comptage : dans les télécommunications par exemple pour compter le nombre de communications dans un intervalle de temps donné, en assurance pour évaluer le nombre de sinistres annuels, en biologie pour évaluer par exemple le nombre d'oeufs par ponte d'un insecte nuisible...

Usage


d<-simulepoisson(n,m)

Arguments

n 	 le nombre de simulations

m 	 la moyenne des simulations

Examples


simulepoisson(1000,4)



[Package lycee version 1.0 Index]