tuyaux {lycee}	R Documentation

tuyaux/controle qualité

Description

On étudie la résistance à la corrosion de différents tuyaux. On dispose de deux variables qualitatives pour tenter d'expliquer la variabilité des observations: le type de protection (peinture) et le type de sol dans lequel ils se trouvent.

Usage

data(tuyaux)

Format

Un tableaux de 27 lignes et 3 variables.

corrosion

une mesure quantitative

sol

le type du sol , 1,2 ou 3

protection

le type de peinture, 1, 2 ou 3

Examples


data(tuyaux)
lire(tuyaux)
boitesparalleles(corrosion,sol)

[Package lycee version 1.0 Index]